requirements.txt
attrs==19.3.0
numpy==1.17.3
Pillow==6.2.1
scipy==1.3.1
